rename: aws_sso => aws_sso_device_code
in order to support grant_type=code and PKCE

following https://aws.amazon.com/about-aws/whats-new/2024/11/aws-command-line-interface-pkce-single-sign-on/
