+++ src/auth_flow_manager.rs
@@ -6,7 +6,7 @@ pub enum AuthFlow {
     Nop,
     OAuthCode(crate::oauth_code::OAuthCodeFlow),
     OAuthDeviceCode(crate::oauth_device_code::OAuthDeviceCodeFlow),
-    AwsSsoDevice(crate::oauth_awssso::AwsSsoDeviceFlow),
+    AwsSsoDevice(crate::oauth_awssso_device_code::AwsSsoDeviceFlow),
 }
 
 impl AuthFlow {
