+++ src/lib.rs
@@ -14,7 +14,7 @@ pub mod proto;
 pub mod token;
 pub mod utils;
 
-pub mod oauth_awssso;
+pub mod oauth_awssso_device_code;
 pub mod oauth_code;
 pub mod oauth_device_code;
 pub mod oauth_refresh_token;
